backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.55.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.12.3
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
deprecation==2.1.0
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.115.14
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.7.0
gotrue==2.12.3
groq==0.29.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.6.4
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
ollama==0.5.1
openai==1.98.0
openai-whisper @ git+https://github.com/openai/whisper.git@c0d2f624c09dc18e709e37c2ad90c039a4eb72a2
packaging==25.0
pillow==11.3.0
postgrest==1.1.1
psutil==6.1.1
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.1
python-multipart==0.0.20
pytesseract==0.3.13
realtime==2.6.0
regex==2025.7.34
requests==2.32.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
storage3==0.12.0
StrEnum==0.4.15
supabase==2.17.0
supafunc==0.10.1
sympy==1.14.0
tiktoken==0.11.0
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
yarl==1.20.1
yt-dlp==2025.6.25